% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{runIR}
\alias{runIR}
\title{Shiny app with exploratory analysis of the term structure of interest rates}
\usage{
runIR()
}
\description{
Shiny app with exploratory analysis of the term structure of interest rates
}
