% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{term_structure}
\alias{term_structure}
\alias{runIR}
\title{Shiny app with exploratory analysis of the term structure of interest rates}
\usage{
runIR()
}
\value{
Runs the Shiny app retreiving the data, running the analysis and showing the report with interactive visualisations.

Dataset term_structure

This is a dataset containing LIBOR interbank interest rates
}
\description{
The function works with no arguments.
}
\section{term_structure.csv}{
NA
}

